// apps/web/src/api/models/getBudgets200DataItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PayPay Money Diary API
 * PayPay 家計簿アプリケーション API
 * OpenAPI spec version: 1.0.0
 */

export type GetBudgets200DataItem = {
  id: string;
  userId: string;
  /** @nullable */
  categoryId: string | null;
  /** @nullable */
  categoryName: string | null;
  /** @minimum 0 */
  amount: number;
  year: number;
  /**
   * @minimum 1
   * @maximum 12
   */
  month: number;
  createdAt: string;
  updatedAt: string;
};