// apps/web/src/api/models/category.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * PayPay Money Diary API
 * PayPay 家計簿アプリケーション API
 * OpenAPI spec version: 1.0.0
 */

export interface Category {
  /** カテゴリID */
  id: string;
  /** カテゴリ名 */
  name: string;
  /** 色コード */
  color: string;
  /**
   * アイコン
   * @nullable
   */
  icon: string | null;
  /** 表示順 */
  displayOrder: number;
  /** デフォルトカテゴリか */
  isDefault: boolean;
}